+++ DataObjects/Migrations/201506161757460_Initial.Designer.cs
@@ -13,7 +13,7 @@ namespace forCrowd.WealthEconomy.DataObjects.Migrations
         
         string IMigrationMetadata.Id
         {
-            get { return "201505300934175_Initial"; }
+            get { return "201506161757460_Initial"; }
         }
         
         string IMigrationMetadata.Source